+++ src/trx/av/audio.c
@@ -1,4 +1,4 @@
-#include <trx/engine/audio_internal.h>
+#include <trx/av/audio_internal.h>
 
 #include <trx/log.h>
 #include <trx/memory.h>
