+++ src/trx/av/audio_sample.c
@@ -1,4 +1,4 @@
-#include <trx/engine/audio_internal.h>
+#include <trx/av/audio_internal.h>
 
 #include <trx/benchmark.h>
 #include <trx/debug.h>
