+++ src/trx/av/audio_stream.c
@@ -1,4 +1,4 @@
-#include <trx/engine/audio_internal.h>
+#include <trx/av/audio_internal.h>
 
 #include <trx/debug.h>
 #include <trx/filesystem.h>
