+++ src/trx/game/shell/common.c
@@ -1,6 +1,6 @@
+#include <trx/av/audio.h>
 #include <trx/config.h>
 #include <trx/debug.h>
-#include <trx/engine/audio.h>
 #include <trx/game/shell.h>
 #include <trx/log.h>
 #include <trx/memory.h>
