+++ src/trx/game/shell/events.c
@@ -1,6 +1,6 @@
+#include <trx/av/audio.h>
 #include <trx/config.h>
 #include <trx/debug.h>
-#include <trx/engine/audio.h>
 #include <trx/game/console/common.h>
 #include <trx/game/fmv.h>
 #include <trx/game/input/common.h>
