+++ tools/sort_imports
@@ -37,10 +37,10 @@ def main() -> None:
             "trx/log_unknown.c": "trx/log.h",
             "trx/log_linux.c": "trx/log.h",
             "trx/log_windows.c": "trx/log.h",
-            "trx/engine/audio.c": "trx/engine/audio_internal.h",
-            "trx/engine/audio_reverb.c": "trx/engine/audio_internal.h",
-            "trx/engine/audio_sample.c": "trx/engine/audio_internal.h",
-            "trx/engine/audio_stream.c": "trx/engine/audio_internal.h",
+            "trx/av/audio.c": "trx/av/audio_internal.h",
+            "trx/av/audio_reverb.c": "trx/av/audio_internal.h",
+            "trx/av/audio_sample.c": "trx/av/audio_internal.h",
+            "trx/av/audio_stream.c": "trx/av/audio_internal.h",
         },
         fix_map={},
         forced_order=[
